Separate handler mapping from routes to enable reverse routing
Our routing library, bidi, defines the routes as data structures, but
with the RouteProvider wants each route in the data structure to also
have a bidi handler defined. This is difficult, because our handler
functions are often stateful and require the application state in the
current system in order to function. Here I separated these two
concerns and wrote some util functions for bidi in order to generate
the application routes in the format that the RouteProvider requires.