Add POST workflow + writing new facts to SSE stream
This adds a POST resource for suggest-session. This is still very
much WIP, because the session storage in the atom is going to be
replaced by the Crux database, and we are probably going to use a
different mechanism (listener?) to listen to changes in the DB and
then write new facts into the SSE stream. But I wanted to sketch out
the behavior of the endpoint + the SSE stream which is produced so
that maybe next week I can look into how to implement the EventSource on
the frontend and implement the first custom element for updating the UI.