Add component for user-notifications
With this commit I add a section to the UI for user notifications.
This section contains a List with the role=log attribute, which
means that when messages are added to the notifications, these will
be read out to a user who is using a screenreader.

This component listens to the same event as the waiting-queue element
(although I renamed it in session-suggested instead of
session-scheduled, since that seems more correct to me).

I also added a UI status element to the `Up Next` section. This element
has role=status, which means that when a session moves up in the line,
we need to modify this status. Then it will also be read out to a screen
reader -- I'm not sure if this is the best place for it, but I'm leaving
it there for now.

In the future, the Notifications may be a feature only for screenreaders,
but I am leaving it visible for everyone for now.