Put data in Crux
This patch introduces https://opencrux.com/ as our data store. For now
all data are stored in an in-memory node that's destroyed when we stop
the system. Replacing it with a persistent solution should be a matter
of changing Crux configuration.

The Crux-related code lives in spacy.crux. That includes:

  - a component that controls the lifecycle of a Crux node,
  - logic for storing and retrieving events (i.e. our aggregates),
  - subscription to Crux transactions, translating them into facts,
    and sending to a provided core.async channel.

Every time we persist an event in a Crux node we store it as:

  - a document with all facts removed, and
  - all the new facts that we haven't stored yet.

Crux is responsible for keeping track of time. Each transaction has
:crux.db/valid-time. When we retrieve a fact from Crux we set its
:spacy.domain/at to :crux.db/valid-time of the transaction in which
it was created.

The core.async channel instantiated in spacy.system will now carry facts
as specified in spacy.domain. spacy.app is responsible for translating
them into server-sent events (see spacy.app/interpret-fact).

spacy.app is decoupled from spacy.crux. A protocol in spacy.data defines
an abstract data store that's implemented by spacy.crux.Crux.