Use our vocabulary more consistently
Events are identified by their URL slugs. Each event also gets
a :crux.db/id but it's an internal Crux detail. This patch should
make it more consistent.

Additionally, there's no need for a slug or an ID when persisting
an event. For now we're assuming the DB is seeded with an event that
has both a slug and an ID assigned beforehand.