Adds schedule-session function to spacy.domain
This adds the function schedule-session to the domain which takes the
first element from the queue and places it in the schedule. I also
implemented the logic for determining if a session is actually able to
be positioned in the slot where we want to place (it needs to be the
first in the queue, the slot must be a valid one, and the slot must
not be already taken). I also added some unit tests with a test event
to test can-schedule-session? function. I'm not completely happy with
the tests: I feel like there is an opportunity here to use some generative
tests, which don't all use the same event data object. But writing the
unit a la TDD did help me to write the can-schedule-session? function with
all of its different conditions, so I figure that it is better than nothing.