Adds endpoint for schedule-session and actions to trigger it in the UI
This adds the schedule-session endpoint to the app.

I've pulled the
transformation of the current state into the domain namespace because
for the new available-slots property, I want to use the namespaced
keywords for the calculation (and we strip the namespaces from the keywords
before sending them to the UI).

I also renamed some variables which were called `session` to now be called
`event`, since we are referring to the current state of our event, not to
a specific session which has been suggested/scheduled.

After this change, I noticed some mistakes with the UI implementation
(the UI custom element always adds the new sessions to the queue, even
if the UpNext area is empty). But I will fix that in another MR.

This doesn't implement any UI behavior for session-scheduled yet:
When you schedule a session, this will by default redirect you back to
the main page and reload it.