Restructures HTML in view + adds very limited styling
One main thing: Now the session that is "up-next" is no longer separated
from the other entries in the Queue. It is listed directly in the ol list
in the queue, and the first entry now simply receives a special styling
with CSS.

This also adds a first CSS layout + very lightweight styling:
* make the sessions look like post-ITs (orange with drop-shadow)
* basic CSS layout for the page
* add horizontal scrolling container around the table so that it doesn't
expand past the viewport on mobile devices
* very minimal styling for the header + nav links

I originally intended to use named CSS colors for the prototype, but they
just aren't very pretty. Instead, I've taken the colors from Tailwind CSS.