Detach facts from events
Events don't contain facts now. Domain functions that modify events
return an outcome that includes the new state of the event and a
collection of new facts. Both are persisted in Crux.

When fetching an event no facts are retrieved.

The next step will be to add domain errors to outcomes.