Pull yada specific code into handler_util
This renames `bidi_util` to `handler_util` and moves yada specific utility
functions into this class. The goal is that the functions in `spacy.app` are
decoupled from the handler implemenation which creates the resources.

For instance, I've created a `handler-util/command` function, which tries
to unify how we execute a command (e.g. execute a command, if it has an error,
return 400 bad request, and if not, persist the outcome and redirect to the
event).

I've also pulled in helper functions for creating a `get-resource` and for
creating an `sse-stream`