Add a spacy.ui/up-next event to the frontend code
When analysing the problem of how to correctly trigger the "up-next" event,
I realized that this is only an issue with the Frontend and not an issue with
the backend, so instead of implementing this as a fact in the backend (I
honestly just felt like I was messing up our backend domain model by trying
to add that fact to it), I decided to create this exclusively as a frontend
CustomEvent.

The waiting-queue custom element in the UI can calculate the current length
of the queue and fire a specific spacy.ui/app event whenever a new element
appears at the beginning of the list. Then the up-next and bulletin-board
custom elements can listen to this event and update their internal state.