Store Crux logs in a relational database
Prior to this commit we stored Crux logs in memory. Now they're kept
in a database using JDBC. For the time being it's SQLite. I also tested
it locally with Postgres. A corresponding db-spec looks as follows:

    {:dbtype "postgres"
     :dbname "postgres"
     :host "localhost"
     :user "postgres"
     :password "do not store passwords in git"}

Once we end up switching to Postgres in production we might want to give
the following paragraph a careful read:

https://github.com/juxt/aero/blob/743e9bc495425b4a4a7c780f5e4b09f6680b4e7a/README.md#hide-passwords-in-local-private-files

I'm extending clojure.core/print-method for Crux to prevent printing
passwords when logging.