Fix bug with template that needs _both_ a data-slot and to render the title
I'm not completely happy with this workaround, but I wanted to add this
quick fix to repair the null pointer exception that I was getting.

The up-next.status.up-next messages is problematic because we both want to
template the message on the server (replacing "${title}" with current title)
_and_ use the same template on the client with the `data-slot` attribute
This workaround currently allows both. An empty `<span data-slot="title">`
is currently rendered when we render the up-next status on the server, but
that is a tradeoff for right now. The i18n / templating mechanism could
definitely use some work.