Refactoring statuses to use role=region and aria-live=polite
This refactors the user notification system (tested on screenreaders).
The hope is to find a system where user updates _that are actually useful_
are read out at the right time (only when something changed and we want
to inform the user about it).

I didn't really like the old usage of role=log, because the user only is
really interested in a status update once, and doesn't want to go back
and read old entries (this is an assumption. could be wrong about that).

So I've exchanged the role=log section for areas which are role=region
as well as aria-live=polite. I have (tried) to test this with a screenreader,
and it seems to work, but will need more testing in the future.