Working on adding visually hidden information to a table
This adds visually hidden text fields with Room & Time information
into the table schedule. (see #62)

Since this information is now in the DOM, I've also modified the
.session element and the `Choose Slot` actions to reference this
information via the `aria-describedby` attribute in the DOM.

I'm not completely sure yet if this is the best solution. I've tested
it myself, but we need to get a screenreader expert to test it if we
want to be sure that the UX is optimal.