Restructure Session component
This restructures the `.session` component. The reason here is that the
way that the session title is currently formatted within the `<summary>`
elements causes empty strings to appear in the accessibility tree that
are labelled as being an "empty toggle" (#59)

This moves the sponsor out of the `<summary>` into a description list
underneath the `<details>` element and adds some styling to it.