Improve navigation and instruction for scheduling a session
This adds a `Schedule session` link to the session object itself which
is blended in when the session is at the top of the queue (I decided to
blend in the link as a progressive enhancement using CSS, because we
already have CSS rules which target the first element in the list).

This also adds some instructions for how to schedule a session that are
shown when you are next in the queue. I've also added some more specific
instructions which are visually hidden, because I think this is probably
more confusing for screenreader users who cannot see the table with all
of the "choose slot" buttons.

The reason is that it was confusing for screenreader users to understand
how to actually schedule a session #60