Modifications for information structure in HTML table
Fixes #62

For some reason, the buttons for choosing a slot weren't working for
screenreaders, so I had to restructre the HTML around that. Adding
a `<div role="group">` around the form and marking the button explicitly
as being type="submit" seems to have solved that problem (Although I
actually don't know which changed actually fixed it).

One change that I made to the information structure was placing the
actual content (the session or the select slot form) before the label.
This is because for screenreaders, you want to access any _content_
before reading the text label.

Also, in screenreader tests it seems to work better to list the time
before the room AND to remove the work "Time / Zeit" from the text:
the screenreader will already say "10:00 o'clock", so the word "Time"
is redundant. This also makes it possible to figure out the time even
earlier because the Screenreader will not have to read out an
unnecessary word.