Fix missing UI notification when a session is deleted
Fixes #73

To fix this, I not only had to add the missing message to the markup,
but I also had to modify the user-notification custom element to have
a `notify-scheduled` option, because we only want the UI change to
trigger when a _scheduled_ session is deleted, not when a session is
deleted from the queue.

Because our facts still contain the `::room` and `::time` data when
the session was scheduled, I check for the existence of these
attributes to implement this option.