Update UI elements for scheduled slots
The old implementation has the issue that the current state that is
shown via the event-slug/ resource can be slightly out of date
compared to the server sent events that are triggered directly
from the Crux transaction log. This means that the original mechanism
for updating the UI wasn't always consistent -- it was possible for
the SSE "session-scheduled" fact to be triggered, and the UI element
would transclude the whole event-slug/ resource and replace the whole
table, but the new scheduled session doesn't appear in the HTML, so
the UI appears to have lost the session.

However, the fact itself contains all of the information that we need
to update the schedule (we don't actually need transclusion, if we
add a bit of logic to the client). What I've done here is create a
`<slot-updater>` custom element which I can embed in each cell of
our board. Then I can write the logic so that each cell can modify
itself when necessary.